Moved `AttractionBalancingZones` to `zoning` module
Renamed class to `BalancingZones` and moved to `zoning` module because
the class is needed by `DVector`.
